requirements.txt
autopep8==1.3.2
boto3==1.4.6
botocore==1.6.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
docutils==0.14
idna==2.5
jmespath==0.9.3
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
pycodestyle==2.3.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.2
s3transfer==0.1.10
six==1.10.0
urllib3==1.22
pydub==0.20.0
